refactor: Optionスクリプトをオプションパネルの表示状態の管理スクリプトと音量設定の変更スクリプトに分割
