+++ Assets/Scenes/Title.unity
@@ -698,8 +698,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1574037441}
-        m_TargetAssemblyTypeName: Option, Assembly-CSharp
+      - m_Target: {fileID: 1574037442}
+        m_TargetAssemblyTypeName: OptionPanelToggle, Assembly-CSharp
         m_MethodName: CloseOption
         m_Mode: 1
         m_Arguments:
@@ -1218,8 +1218,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1574037441}
-        m_TargetAssemblyTypeName: Option, Assembly-CSharp
+      - m_Target: {fileID: 1574037442}
+        m_TargetAssemblyTypeName: OptionPanelToggle, Assembly-CSharp
         m_MethodName: OpenOption
         m_Mode: 1
         m_Arguments:
@@ -1301,7 +1301,7 @@ RectTransform:
   m_Father: {fileID: 957886375}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -1902,7 +1902,7 @@ RectTransform:
   m_Father: {fileID: 1123403973}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -1961,7 +1961,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1511031887
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2170,6 +2170,7 @@ GameObject:
   - component: {fileID: 1574037439}
   - component: {fileID: 1574037438}
   - component: {fileID: 1574037437}
+  - component: {fileID: 1574037442}
   - component: {fileID: 1574037441}
   m_Layer: 5
   m_Name: Canvas
@@ -2276,12 +2277,24 @@ MonoBehaviour:
   m_GameObject: {fileID: 1574037436}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ecdfa58176d45ec4e9a23d467f070353, type: 3}
+  m_Script: {fileID: 11500000, guid: dc136e7e2b8efca4d9a363ee36696288, type: 3}
   m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::Option
-  optionPanel: {fileID: 1511031886}
+  m_EditorClassIdentifier: Assembly-CSharp::VolumeController
   bgmSlider: {fileID: 1809860709}
   seSlider: {fileID: 647241861}
+--- !u!114 &1574037442
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1574037436}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9df2f871069b09646823129177053ac4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: Assembly-CSharp::OptionPanelToggle
+  optionPanel: {fileID: 1511031886}
 --- !u!1 &1790101120
 GameObject:
   m_ObjectHideFlags: 0
@@ -2481,7 +2494,7 @@ RectTransform:
   m_Father: {fileID: 2124224117}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 10, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -2795,7 +2808,7 @@ RectTransform:
   m_Father: {fileID: 1298149701}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 20, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
