--- Assets/Scripts/Title/Option.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ecdfa58176d45ec4e9a23d467f070353
